requirement.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
decorator==5.1.1
Django==5.0.6
django-cors-headers==4.3.1
django-debug-toolbar==4.4.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
executing==2.0.1
filelock==3.14.0
fsspec==2024.5.0
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
huggingface-hub==0.23.2
idna==3.7
ipython==8.24.0
jedi==0.19.1
matplotlib-inline==0.1.7
mysql-client==0.0.1
mysql-connector==2.2.9
mysql-connector-python==8.4.0
numpy==1.26.4
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PySocks==1.7.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
requests-file==2.1.0
safetensors==0.4.3
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tldextract==5.1.2
tokenizers==0.19.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
typing_extensions==4.12.1
urllib3==2.2.1
uv==0.2.5
wcwidth==0.2.13
zope.event==5.0
zope.interface==6.4.post2
